+++ R/pulse_shapes.R
@@ -11,3 +11,22 @@ get_guassian_pulse <- function(angle, n, trunc = 1) {
   B <- B / sum(B) * pi * angle / 180
   B
 }
+
+#' Read a .pta formatted text file compatible with Siemens PulseTool.
+#' @param fname pta formatted pulse file name.
+#' @return pulse waveform and header.
+#' @export
+read_pulse_pta <- function(fname) {
+  con <- file(fname, "r")
+  header <- utils::read.table(con, nrows = 8, sep = ":", strip.white = TRUE)
+  suppressWarnings(header$V3 <- as.numeric(header$V2))
+  is_txt <- is.na(header$V3)
+  header_txt <- stats::setNames(as.list(header$V2[is_txt]), header$V1[is_txt])
+  header_num <- stats::setNames(as.list(header$V3[!is_txt]), header$V1[!is_txt])
+  header_out <- c(header_txt, header_num)
+  pulse  <- read.table(con)
+  close(con)
+  pulse  <- pulse[, 1:2]
+  colnames(pulse) <- c("mag", "pha")
+  return(list(data = pulse, header = header_out))
+}